refactor: menu, menu group 이동
