+++ src/test/java/kitchenpos/Fixtures.java
@@ -1,6 +1,8 @@
 package kitchenpos;
 
 import kitchenpos.domain.*;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.product.domain.Product;
 
 import java.math.BigDecimal;
