+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -1,7 +1,14 @@
-package kitchenpos.application;
+package kitchenpos.menu.application;
 
+import kitchenpos.application.FakePurgomalumClient;
 import kitchenpos.domain.*;
 import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.menu.application.MenuService;
+import kitchenpos.menu.domain.InMemoryMenuGroupRepository;
+import kitchenpos.menu.domain.InMemoryMenuRepository;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.InMemoryProductRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
