+++ src/test/java/kitchenpos/menu/domain/InMemoryMenuGroupRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.application;
+package kitchenpos.menu.domain;
 
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroupRepository;
 
 import java.util.*;
