+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -1,10 +1,9 @@
 package kitchenpos.product.application;
 
 import kitchenpos.application.FakePurgomalumClient;
-import kitchenpos.application.InMemoryMenuRepository;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuRepository;
-import kitchenpos.product.application.ProductService;
+import kitchenpos.menu.domain.InMemoryMenuRepository;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.InMemoryProductRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
